pyproject.toml
[tool.poetry]
name = "serverless-and-ai"
version = "0.1.0"
description = "This is the project showcasing capabilities of AI working together in a serverless manner within AWS cloud."
authors = ["amswiatkowski <adam.swiatkowski@iodi.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

# 3rdparties
boto3 = "^1.28.62"
botocore = "^1.28.62"
aws-lambda-context = "*"
pydantic = "*"
aws-lambda-powertools =  {extras = ["all"],version = "*"}
requests = "*"
cachetools = "*"
boto3-stubs = { version = "*", extras = ["cloudformation", "bedrock", "lambda"] }
aws-lambda-typing = "*"
langchain = "*"
langchain-community = "*"
langchain-aws = "*"
redis = "*"

[tool.poetry.group.dev.dependencies]
# Internal
serverless-and-ai-cdk = {path = "cdk", develop = true}

# 3rdparties
pytest = "*"
pytest-mock = "*"
pytest-dotenv = "*"
pycodestyle = "*"
python-dotenv = "*"
GitPython = "*"
yapf = "*"
pylint = "*"
isort = "*"
"aws-cdk.aws-lambda-python-alpha" = "*"
streamlit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"